change pickle

+++ src/utils/cnn.py
@@ -28,8 +28,8 @@ num_train, depth, height, width = X_train.shape # there are 50000 training examp
 num_test = X_test.shape[0] # there are 10000 test examples in CIFAR-10
 num_classes = np.unique(y_train).shape[0] # there are 10 image classes
 
-X_train = X_train.astype('float32')
-X_test = X_test.astype('float32')
+X_train = X_train.astype('float64')
+X_test = X_test.astype('float64')
 X_train /= np.max(X_train) # Normalise data to [0, 1] range
 X_test /= np.max(X_test) # Normalise data to [0, 1] range
